[#22] Move Array out of system/base.py
The Array type hint was causing a circular import.
