+++ src/stanshock/system/backend.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+import numpy as np
+
+# Define the Array type for use in type hints to make future changes easier
+Array = np.ndarray
