+++ src/stanshock/system/base.py
@@ -2,12 +2,8 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 
-import numpy as np
-
 from stanshock.physics.fluid_base import FluidPhysics, FluidState
-
-# Define the Array type for use in type hints to make future changes easier
-Array = np.ndarray
+from stanshock.system.backend import Array
 
 
 class RightHandSide(ABC):
